refactor: 컨트롤러 네이밍 수정
